requirements.txt
pyyaml
discord.py
python-twitter
